requirements.txt
streamlit==1.52.2
openai==2.14.0
python-dotenv==1.2.1
PyPDF2==3.0.1
annoy==1.17.1